Turn options check into a function.
